+++ src/components/linkedin/utilities.ts
@@ -1,16 +1,14 @@
 import { createScriptLoader } from '@solid-primitives/script-loader';
 
-let isLinkedInScriptAdded = false;
-export const linkedInClasses = [`.LI-profile-badge`, `.LI-simple-link`].join(`,`);
-const linkedInBadgesUrl = `https://platform.linkedin.com/badges/js/profile.js`;
+export const linkedInClasses = ['.LI-profile-badge', '.LI-simple-link'].join(',');
+const linkedInBadgesUrl = 'https://platform.linkedin.com/badges/js/profile.js';
 
 const LIRenderAll = (): void | unknown =>
-  window.LI && typeof window.LIRenderAll === `function` && window.LIRenderAll();
+  window.LI && typeof window.LIRenderAll === 'function' && window.LIRenderAll();
 
 export const handleLinkedInLoad = (): { status: 'createScriptLoader' | 'LILoad' } => {
-  if (document.querySelector(linkedInClasses) && !isLinkedInScriptAdded) {
+  if (document.querySelector(linkedInClasses) && !(window.LI && window.LIRenderAll)) {
     createScriptLoader({ src: linkedInBadgesUrl });
-    isLinkedInScriptAdded = true;
     return {
       status: 'createScriptLoader',
     };