+++ src/components/listennotes/index.tsx
@@ -1,0 +1,1 @@
+export { ListenNotesEpisode } from 'components/listennotes/listennotes-episode';