+++ src/components/twitter/utilities.ts
@@ -1,22 +1,19 @@
 import { createScriptLoader } from '@solid-primitives/script-loader';
 
-let isTwttrScriptAdded = false;
-
 export const twttrClasses = [
-  `.twitter-tweet`,
-  `.twitter-timeline`,
-  `.twitter-follow-button`,
-  `.twitter-mention-button`,
-  `.twitter-hashtag-button`,
-].join(`,`);
+  '.twitter-tweet',
+  '.twitter-timeline',
+  '.twitter-follow-button',
+  '.twitter-mention-button',
+  '.twitter-hashtag-button',
+].join(',');
 
 const twttrLoad = (): void =>
-  window.twttr?.widgets?.load?.(document.querySelectorAll(`solid-social`));
+  window.twttr?.widgets?.load?.(document.querySelectorAll(twttrClasses));
 
 export const handleTwttrLoad = (): { status: 'createScriptLoader' | 'twttrLoad' } => {
-  if (document.querySelector(twttrClasses) && !isTwttrScriptAdded) {
+  if (document.querySelector(twttrClasses) && !window.twttr) {
     createScriptLoader({ src: '//platform.twitter.com/widgets.js' });
-    isTwttrScriptAdded = true;
     return {
       status: 'createScriptLoader',
     };