+++ src/components/wistia/utilities.ts
@@ -1,12 +1,11 @@
+import { createSignal } from 'solid-js';
 import { createScriptLoader } from '@solid-primitives/script-loader';
 
-let isWistiaAdded = false;
+const [isWistiaAdded, setWistiaAdded] = createSignal(false);
 const embedUrl = `//fast.wistia.net/assets/external/E-v1.js`;
 
 export const handleWistiaLoad = (): void => {
-  if (isWistiaAdded) {
-    return;
-  }
+  if (isWistiaAdded()) return;
   createScriptLoader({ src: embedUrl });
-  isWistiaAdded = true;
+  setWistiaAdded(true);
 };