+++ src/index.ts
@@ -10,6 +10,7 @@ export { Gist } from 'components/gist';
 export { Instagram } from 'components/instagram';
 export { Lbry } from 'components/lbry';
 export { LinkedInBadge } from 'components/linkedin';
+export { ListenNotesEpisode } from 'components/listennotes';
 export { Pin, PinterestBoard, PinterestFollowButton } from 'components/pinterest';
 export { SimplecastEpisode } from 'components/simplecast';
 export { Snack } from 'components/snack';